+++ game/scrabble.py
@@ -73,17 +73,19 @@ class ScrabbleGame:
 
     def validateWord(self, word, location, orientation):
         (x, y) = location
-        for _ in word:
-            if self.current_player.hasWord(word) and self.board.validate_word_inside_board(word, location, orientation):
-                word = unidecode(word)
-                return True;
-            elif (
-                self.isWordInBoard(word, location, orientation)
-                and
-                self.board.validate_word_inside_board(word, location, orientation)
-            ):
-                word = unidecode(word)
-                return True;
+        if self.current_player.hasWord(word) and self.board.validate_word_inside_board(word, location, orientation) and dict(word):
+            word = unidecode(word)
+            return True;
+        elif (
+            self.isWordInBoard(word, location, orientation)
+            and
+            self.board.validate_word_inside_board(word, location, orientation)
+            and 
+            dict(word)
+        ):
+            word = unidecode(word)
+            return True;
+            
 
     def putWord(self, word, location, orientation):
         (x, y) = location